# src/strada/ride-service/src/ride-request/controllers/ride-request.controller.ts
import {
  Controller,
  Get,
  Post,
  Put, Body,
  Param,
  Query,
  HttpStatus,
  HttpCode,
  ValidationPipe
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiParam,
  ApiQuery,
  ApiBearerAuth,
  ApiBody,
} from '@nestjs/swagger';
import { RideRequestService } from '../services/ride-request.service';
import { RequestStatus } from '@prisma/client';

// DTOs para validação de requisições
class LocationDto {
  address: string;
  latitude: number;
  longitude: number;
}

class CreateRideRequestDto {
  rideId: string;
  passengerId: string;
  seatsNeeded: number;
  message?: string;
  pickupLocation?: LocationDto;
  dropoffLocation?: LocationDto;
  preferredDepartureTime?: Date;
  maxWalkingDistance?: number;
  maxDetourTime?: number;
}

class SearchNearbyDto {
  latitude: number;
  longitude: number;
  radiusMeters?: number;
  status?: string;
  limit?: number;
  offset?: number;
}

class RouteSearchDto {
  routeGeometry: string;
  maxDistance?: number;
  status?: string;
  limit?: number;
  offset?: number;
}

class OptimalPickupDto {
  passengerStart: {
    latitude: number;
    longitude: number;
  };
  passengerEnd: {
    latitude: number;
    longitude: number;
  };
  maxDetourKm?: number;
}

class BulkStatusUpdateDto {
  requestIds: string[];
  newStatus: string;
  reason?: string;
}

class FindCompatibleRidesDto {
  pickupLocation: LocationDto;
  dropoffLocation: LocationDto;
  departureTime?: Date;
  maxWalkingDistance?: number;
  maxDetourTime?: number;
  seatsNeeded?: number;
}

@ApiTags('Ride Requests')
@Controller('rides')
@ApiBearerAuth()
export class RideRequestController {
  constructor(private readonly rideRequestService: RideRequestService) {}

  // ========== BASIC CRUD OPERATIONS ==========

  @Post(':rideId/requests')
  @HttpCode(HttpStatus.CREATED)
  @ApiOperation({
    summary:
      'Request to join a ride with optional custom pickup/dropoff locations',
  })
  @ApiParam({ name: 'rideId', description: 'UUID of the ride' })
  @ApiBody({ type: CreateRideRequestDto })
  @ApiResponse({
    status: 201,
    description: 'Ride request created successfully with detour analysis',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid request data or excessive detour',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride not found',
  })
  async createRideRequest(
    @Param('rideId') rideId: string,
    @Body(ValidationPipe) createRideRequestDto: CreateRideRequestDto,
  ) {
    return await this.rideRequestService.createRideRequest({
      ...createRideRequestDto,
      rideId,
    });
  }

  @Get('requests/:id')
  @ApiOperation({ summary: 'Get ride request details by ID' })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description: 'Ride request details retrieved successfully',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async getRideRequest(@Param('id') id: string) {
    return await this.rideRequestService.findRideRequestById(id);
  }

  @Get(':rideId/requests')
  @ApiOperation({ summary: 'Get all requests for a specific ride' })
  @ApiParam({ name: 'rideId', description: 'UUID of the ride' })
  @ApiQuery({
    name: 'status',
    required: false,
    description:
      'Filter by request status (PENDING, ACCEPTED, REJECTED, CANCELLED)',
  })
  @ApiResponse({
    status: 200,
    description: 'List of ride requests retrieved successfully',
  })
  async getRideRequests(
    @Param('rideId') rideId: string,
    @Query('status') status?: string,
  ) {
    return await this.rideRequestService.findRideRequestsByRideId(rideId);
    // Observação: O serviço não suporta filtro por status diretamente. Se necessário, adicione lógica no serviço.
  }

  @Get('passengers/:passengerId/requests')
  @ApiOperation({ summary: 'Get all requests made by a specific passenger' })
  @ApiParam({ name: 'passengerId', description: 'UUID of the passenger' })
  @ApiQuery({
    name: 'status',
    required: false,
    description: 'Filter by request status',
  })
  @ApiResponse({
    status: 200,
    description: 'List of passenger ride requests retrieved successfully',
  })
  async getPassengerRequests(
    @Param('passengerId') passengerId: string,
    @Query('status') status?: string,
  ) {
    return await this.rideRequestService.findRideRequestsByPassengerId(
      passengerId,
      status,
    );
  }

  // ========== REQUEST STATUS MANAGEMENT ==========

  @Put('requests/:id/accept')
  @ApiOperation({ summary: 'Accept a ride request (Driver action)' })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description: 'Ride request accepted successfully. Available seats updated.',
  })
  @ApiResponse({
    status: 400,
    description: 'Not enough available seats or request cannot be accepted',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async acceptRideRequest(@Param('id') id: string) {
    return await this.rideRequestService.acceptRideRequest(id);
  }

  @Put('requests/:id/reject')
  @ApiOperation({ summary: 'Reject a ride request (Driver action)' })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiBody({
    schema: {
      type: 'object',
      properties: {
        reason: {
          type: 'string',
          description: 'Optional reason for rejection',
        },
      },
    },
  })
  @ApiResponse({
    status: 200,
    description: 'Ride request rejected successfully',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async rejectRideRequest(
    @Param('id') id: string,
    @Body('reason') reason?: string,
  ) {
    // Observação: O método rejectRideRequest no serviço não aceita um parâmetro 'reason'.
    // Se necessário, modifique o serviço para suportar 'reason'.
    return await this.rideRequestService.rejectRideRequest(id);
  }

  @Put('requests/:id/cancel')
  @ApiOperation({ summary: 'Cancel a ride request (Passenger action)' })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description:
      'Ride request cancelled successfully. Seats returned if previously accepted.',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async cancelRideRequest(@Param('id') id: string) {
    return await this.rideRequestService.cancelRideRequest(id);
  }

  @Put('requests/bulk-update-status')
  @ApiOperation({
    summary: 'Bulk update status of multiple ride requests (Admin function)',
  })
  @ApiBody({ type: BulkStatusUpdateDto })
  @ApiResponse({
    status: 200,
    description: 'Bulk status update completed with success/failure summary',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid status or malformed request',
  })
  async bulkUpdateRequestStatus(
    @Body(ValidationPipe) updateDto: BulkStatusUpdateDto,
  ) {
    return await this.rideRequestService.bulkUpdateStatus(
      updateDto.requestIds,
      updateDto.newStatus as RequestStatus,
    );
    // Observação: O serviço não usa o campo 'reason'. Se necessário, modifique o serviço.
  }

  // ========== GEOGRAPHIC SEARCH ==========

  @Get('requests/search/nearby')
  @ApiOperation({
    summary:
      'Search for ride requests within a geographic radius (PostGIS-powered)',
  })
  @ApiQuery({ name: 'centerLat', description: 'Center latitude for search' })
  @ApiQuery({ name: 'centerLng', description: 'Center longitude for search' })
  @ApiQuery({
    name: 'radiusMeters',
    required: false,
    description: 'Search radius in meters (default: 5000, max: 50000)',
  })
  @ApiQuery({
    name: 'status',
    required: false,
    description: 'Filter by request status',
  })
  @ApiQuery({
    name: 'limit',
    required: false,
    description: 'Maximum number of results (default: 20)',
  })
  @ApiQuery({
    name: 'offset',
    required: false,
    description: 'Number of results to skip for pagination',
  })
  @ApiResponse({
    status: 200,
    description:
      'Ride requests within radius retrieved with distance calculations',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid coordinates or search radius too large',
  })
  async searchNearbyRequests(
    @Query(ValidationPipe) searchParams: SearchNearbyDto,
  ) {
    return await this.rideRequestService.getRequestsWithinRadius({
      latitude: searchParams.latitude,
      longitude: searchParams.longitude,
      radiusMeters: searchParams.radiusMeters,
    });
    // Observação: O serviço não suporta 'status', 'limit' ou 'offset'. Adicione suporte no serviço se necessário.
  }

  @Post('requests/search/by-route')
  @ApiOperation({
    summary:
      'Find ride requests along a specific route geometry (PostGIS-powered)',
  })
  @ApiBody({ type: RouteSearchDto })
  @ApiResponse({
    status: 200,
    description:
      'Ride requests along route retrieved with route distance calculations',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid route geometry or distance parameters',
  })
  async searchRequestsByRoute(
    @Body(ValidationPipe) searchParams: RouteSearchDto,
  ) {
    return await this.rideRequestService.getRequestsByRoute({
      routeGeometry: searchParams.routeGeometry,
      maxDistance: searchParams.maxDistance,
    });
    // Observação: O serviço não suporta 'status', 'limit' ou 'offset'. Adicione suporte no serviço se necessário.
  }

  // ========== SMART MATCHING & OPTIMIZATION ==========

  @Post(':rideId/requests/optimal-pickup')
  @ApiOperation({
    summary:
      'Find optimal pickup and dropoff points along ride route (PostGIS spatial analysis)',
  })
  @ApiParam({ name: 'rideId', description: 'UUID of the ride' })
  @ApiBody({ type: OptimalPickupDto })
  @ApiResponse({
    status: 200,
    description:
      'Optimal pickup/dropoff points calculated with walking distances and detour analysis',
  })
  @ApiResponse({
    status: 400,
    description:
      'Invalid passenger coordinates or excessive detour requirement',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride not found',
  })
  async findOptimalPickupDropoff(
    @Param('rideId') rideId: string,
    @Body(ValidationPipe) params: OptimalPickupDto,
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // findOptimalPickupDropoff(rideId: string, params: OptimalPickupDto)
    throw new Error('Método não implementado no serviço');
  }

  @Post('passengers/:passengerId/find-compatible-rides')
  @ApiOperation({
    summary: 'Find compatible rides for a passenger based on their preferences',
  })
  @ApiParam({ name: 'passengerId', description: 'UUID of the passenger' })
  @ApiBody({ type: FindCompatibleRidesDto })
  @ApiResponse({
    status: 200,
    description:
      'Compatible rides found with match scoring and detour analysis',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid location data or search parameters',
  })
  async findCompatibleRides(
    @Param('passengerId') passengerId: string,
    @Body(ValidationPipe) params: FindCompatibleRidesDto,
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // findCompatibleRides(passengerId: string, params: FindCompatibleRidesDto)
    throw new Error('Método não implementado no serviço');
  }

  @Get('requests/:id/smart-recommendations')
  @ApiOperation({
    summary: 'Get AI-powered ride recommendations for a specific request',
  })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description:
      'Smart recommendations with compatibility scoring and historical data',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async getSmartRideRecommendations(@Param('id') id: string) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getSmartRideRecommendations(id: string)
    throw new Error('Método não implementado no serviço');
  }

  @Get('requests/:id/priority-score')
  @ApiOperation({
    summary:
      'Calculate priority score for a ride request based on multiple factors',
  })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description: 'Priority score calculated with detailed factor analysis',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async calculateRequestPriority(@Param('id') id: string) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // calculateRequestPriority(id: string)
    throw new Error('Método não implementado no serviço');
  }

  // ========== ANALYTICS & STATISTICS ==========

  @Get('requests/statistics')
  @ApiOperation({
    summary: 'Get ride request statistics and analytics',
  })
  @ApiQuery({
    name: 'rideId',
    required: false,
    description: 'Filter statistics by ride ID',
  })
  @ApiQuery({
    name: 'passengerId',
    required: false,
    description: 'Filter statistics by passenger ID',
  })
  @ApiQuery({
    name: 'dateFrom',
    required: false,
    description: 'Start date for statistics (ISO format)',
  })
  @ApiQuery({
    name: 'dateTo',
    required: false,
    description: 'End date for statistics (ISO format)',
  })
  @ApiResponse({
    status: 200,
    description:
      'Ride request statistics with acceptance rates and response times',
  })
  async getRideRequestStatistics(
    @Query('rideId') rideId?: string,
    @Query('passengerId') passengerId?: string,
    @Query('dateFrom') dateFrom?: string,
    @Query('dateTo') dateTo?: string,
  ) {
    // Observação: O serviço suporta apenas rideId e passengerId. Filtros de data precisam ser adicionados ao método getRequestStatistics.
    return await this.rideRequestService.getRequestStatistics({
      rideId,
      passengerId,
    });
  }

  @Get('requests/analytics/density-heatmap')
  @ApiOperation({
    summary:
      'Get ride request density heatmap data for geographic visualization',
  })
  @ApiQuery({
    name: 'bounds',
    description: 'Bounding box in format: minLng,minLat,maxLng,maxLat',
  })
  @ApiQuery({
    name: 'gridSize',
    required: false,
    description: 'Grid cell size in meters (default: 1000)',
  })
  @ApiQuery({
    name: 'dateFrom',
    required: false,
    description: 'Start date for analysis (ISO format)',
  })
  @ApiQuery({
    name: 'dateTo',
    required: false,
    description: 'End date for analysis (ISO format)',
  })
  @ApiResponse({
    status: 200,
    description: 'Heatmap data with ride request density by geographic grid',
  })
  @ApiResponse({
    status: 400,
    description: 'Invalid bounds format or date parameters',
  })
  async getRequestDensityHeatmap(
    @Query('bounds') bounds: string,
    @Query('gridSize') gridSize?: number,
    @Query('dateFrom') dateFrom?: string,
    @Query('dateTo') dateTo?: string,
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getRequestDensityHeatmap(params: { bounds: string, gridSize?: number, dateFrom?: string, dateTo?: string })
    throw new Error('Método não implementado no serviço');
  }

  @Get('requests/analytics/popular-routes')
  @ApiOperation({
    summary: 'Get popular pickup/dropoff route combinations',
  })
  @ApiQuery({
    name: 'startLat',
    required: false,
    description: 'Filter routes starting near this latitude',
  })
  @ApiQuery({
    name: 'startLng',
    required: false,
    description: 'Filter routes starting near this longitude',
  })
  @ApiQuery({
    name: 'radius',
    required: false,
    description: 'Search radius in meters (default: 5000)',
  })
  @ApiQuery({
    name: 'limit',
    required: false,
    description: 'Maximum number of popular routes to return (default: 10)',
  })
  @ApiQuery({
    name: 'dateFrom',
    required: false,
    description: 'Start date for route analysis (ISO format)',
  })
  @ApiQuery({
    name: 'dateTo',
    required: false,
    description: 'End date for route analysis (ISO format)',
  })
  @ApiResponse({
    status: 200,
    description: 'List of popular route combinations with usage statistics',
  })
  async getPopularRoutes(
    @Query('startLat') startLat?: number,
    @Query('startLng') startLng?: number,
    @Query('radius') radius?: number,
    @Query('limit') limit?: number,
    @Query('dateFrom') dateFrom?: string,
    @Query('dateTo') dateTo?: string,
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getPopularRoutes(params: { startLat?: number, startLng?: number, radius?: number, limit?: number, dateFrom?: string, dateTo?: string })
    throw new Error('Método não implementado no serviço');
  }

  @Get('requests/analytics/performance-metrics')
  @ApiOperation({
    summary: 'Get comprehensive performance metrics for ride requests',
  })
  @ApiQuery({
    name: 'timeframe',
    required: false,
    description:
      'Timeframe for analysis: daily, weekly, monthly (default: weekly)',
  })
  @ApiQuery({
    name: 'includeGeographic',
    required: false,
    description: 'Include geographic breakdown (default: false)',
  })
  @ApiResponse({
    status: 200,
    description: 'Comprehensive performance metrics and trends',
  })
  async getPerformanceMetrics(
    @Query('timeframe') timeframe?: string,
    @Query('includeGeographic') includeGeographic?: boolean,
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getPerformanceMetrics(params: { timeframe?: string, includeGeographic?: boolean })
    throw new Error('Método não implementado no serviço');
  }

  // ========== REAL-TIME FEATURES ==========

  @Get('requests/live/dashboard')
  @ApiOperation({
    summary: 'Get real-time dashboard data for ride requests',
  })
  @ApiResponse({
    status: 200,
    description: 'Live dashboard data with current statistics',
  })
  async getLiveDashboard() {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getLiveDashboardData()
    throw new Error('Método não implementado no serviço');
  }

  @Get('requests/:id/live-status')
  @ApiOperation({
    summary: 'Get real-time status updates for a specific ride request',
  })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description: 'Real-time status information with live updates',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async getLiveRequestStatus(@Param('id') id: string) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // getLiveRequestStatus(id: string)
    throw new Error('Método não implementado no serviço');
  }

  @Post('requests/:id/notify-alternatives')
  @ApiOperation({
    summary:
      'Send alternative ride suggestions to passenger if request is taking too long',
  })
  @ApiParam({ name: 'id', description: 'UUID of the ride request' })
  @ApiResponse({
    status: 200,
    description: 'Alternative suggestions sent successfully',
  })
  @ApiResponse({
    status: 404,
    description: 'Ride request not found',
  })
  async notifyAlternatives(@Param('id') id: string) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // notifyAlternativeRides(id: string)
    throw new Error('Método não implementado no serviço');
  }

  // ========== ADMIN UTILITIES ==========

  @Get('requests/admin/health-check')
  @ApiOperation({
    summary: 'Health check for ride request system (Admin only)',
  })
  @ApiResponse({
    status: 200,
    description: 'System health status with detailed metrics',
  })
  async adminHealthCheck() {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // performHealthCheck()
    throw new Error('Método não implementado no serviço');
  }

  @Post('requests/admin/cleanup-stale')
  @ApiOperation({
    summary: 'Clean up stale ride requests (Admin only)',
  })
  @ApiBody({
    schema: {
      type: 'object',
      properties: {
        olderThanDays: {
          type: 'number',
          description: 'Remove requests older than X days',
        },
        dryRun: {
          type: 'boolean',
          description: 'Preview changes without executing',
        },
      },
    },
  })
  @ApiResponse({
    status: 200,
    description: 'Cleanup completed with summary of actions taken',
  })
  async cleanupStaleRequests(
    @Body() body: { olderThanDays?: number; dryRun?: boolean },
  ) {
    // Observação: Não há método correspondente no serviço. Sugiro adicionar um método como:
    // cleanupStaleRequests(olderThanDays: number, dryRun: boolean)
    throw new Error('Método não implementado no serviço');
  }
}