# src/strada/ride-service/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rides": {
      "post": {
        "operationId": "RideController_createRide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ride created successfully"
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "summary": "Create a new ride",
        "tags": [
          "rides"
        ]
      }
    },
...
    "/rides/{rideId}/requests": {
      "post": {
        "operationId": "RideRequestController_createRideRequest",
        "parameters": [
          {
            "name": "rideId",
            "required": true,
            "in": "path",
            "description": "UUID of the ride",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ride request created successfully with detour analysis"
          },
          "400": {
            "description": "Invalid request data or excessive detour"
          },
          "404": {
            "description": "Ride not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request to join a ride with optional custom pickup/dropoff locations",
        "tags": [
          "Ride Requests"
        ]
      },
      "get": {
        "operationId": "RideRequestController_getRideRequests",
        "parameters": [
          {
            "name": "rideId",
            "required": true,
            "in": "path",
            "description": "UUID of the ride",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by request status (PENDING, ACCEPTED, REJECTED, CANCELLED)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ride requests retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all requests for a specific ride",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/{id}": {
      "get": {
        "operationId": "RideRequestController_getRideRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of the ride request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ride request details retrieved successfully"
          },
          "404": {
            "description": "Ride request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get ride request details by ID",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/passengers/{passengerId}/requests": {
      "get": {
        "operationId": "RideRequestController_getPassengerRequests",
        "parameters": [
          {
            "name": "passengerId",
            "required": true,
            "in": "path",
            "description": "UUID of the passenger",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by request status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of passenger ride requests retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all requests made by a specific passenger",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/{id}/accept": {
      "put": {
        "operationId": "RideRequestController_acceptRideRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of the ride request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ride request accepted successfully. Available seats updated."
          },
          "400": {
            "description": "Not enough available seats or request cannot be accepted"
          },
          "404": {
            "description": "Ride request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept a ride request (Driver action)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/{id}/reject": {
      "put": {
        "operationId": "RideRequestController_rejectRideRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of the ride request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Optional reason for rejection"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ride request rejected successfully"
          },
          "404": {
            "description": "Ride request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a ride request (Driver action)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/{id}/cancel": {
      "put": {
        "operationId": "RideRequestController_cancelRideRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID of the ride request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ride request cancelled successfully. Seats returned if previously accepted."
          },
          "404": {
            "description": "Ride request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a ride request (Passenger action)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/bulk-update-status": {
      "put": {
        "operationId": "RideRequestController_bulkUpdateRequestStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk status update completed with success/failure summary"
          },
          "400": {
            "description": "Invalid status or malformed request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk update status of multiple ride requests (Admin function)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/search/nearby": {
      "get": {
        "operationId": "RideRequestController_searchNearbyRequests",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip for pagination",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results (default: 20)",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by request status",
            "schema": {}
          },
          {
            "name": "radiusMeters",
            "required": false,
            "in": "query",
            "description": "Search radius in meters (default: 5000, max: 50000)",
            "schema": {}
          },
          {
            "name": "centerLng",
            "required": true,
            "in": "query",
            "description": "Center longitude for search",
            "schema": {}
          },
          {
            "name": "centerLat",
            "required": true,
            "in": "query",
            "description": "Center latitude for search",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Ride requests within radius retrieved with distance calculations"
          },
          "400": {
            "description": "Invalid coordinates or search radius too large"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search for ride requests within a geographic radius (PostGIS-powered)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/requests/search/by-route": {
      "post": {
        "operationId": "RideRequestController_searchRequestsByRoute",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ride requests along route retrieved with route distance calculations"
          },
          "400": {
            "description": "Invalid route geometry or distance parameters"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find ride requests along a specific route geometry (PostGIS-powered)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
    "/rides/{rideId}/requests/optimal-pickup": {
      "post": {
        "operationId": "RideRequestController_findOptimalPickupDropoff",
        "parameters": [
          {
            "name": "rideId",
            "required": true,
            "in": "path",
            "description": "UUID of the ride",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimalPickupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Optimal pickup/dropoff points calculated with walking distances and detour analysis"
          },
          "400": {
            "description": "Invalid passenger coordinates or excessive detour requirement"
          },
          "404": {
            "description": "Ride not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find optimal pickup and dropoff points along ride route (PostGIS spatial analysis)",
        "tags": [
          "Ride Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Strada",
    "description": "A simple API to manage and share rides",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateRideData": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRideRequestDto": {
        "type": "object",
        "properties": {}
      },
      "BulkStatusUpdateDto": {
        "type": "object",
        "properties": {}
      },
      "RouteSearchDto": {
        "type": "object",
        "properties": {}
      },
      "OptimalPickupDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}